requirements-dev.txt
# Development dependencies for crypto-dt-source
# Install with: pip install -r requirements-dev.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Security
bandit==1.7.5
safety==2.3.5

# Type stubs
types-requests==2.31.0.10
types-aiohttp==3.9.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Debugging
ipdb==0.13.13
ipython==8.19.0

# Pre-commit hooks
pre-commit==3.6.0

# Additional JWT support
PyJWT==2.8.0
cryptography==41.0.7
